MAR-911: Alleviate most Airflow concurrency issues by using millisecond accuracy in execution date
The stack trace pointed to an issue in DAGRun.refresh_from_db(),
but the issue was actually a (silenced) session rollback in set_dirty().
This error was related to multiple DAG runs having the same execution
date (since millisecond precision was turned off by default, therefore
limiting the maximum throughpout to 1 dag run per dag id and per
second).
The error is now explicitly raised, and the millisecond accuracy
enforced.